가격정책 구현, 디너 종류/스타일 Enum화 (#18)
* Enum화

* js수정

---------

+++ src/main/java/com/devak/mrdaebakdinner/entity/ItemEntity.java
@@ -15,5 +15,5 @@ public class ItemEntity {
     @Column(nullable = false, unique = true)
     private String name;
     @Column(nullable = false, name = "unit_price")
-    private Double unitPrice; // 단가
+    private Integer unitPrice; // 단가
 }
